Restricted Pages

+++ base/views.py
@@ -1,19 +1,25 @@
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.db.models import Q
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 
 from .forms import RoomForm
 from .models import Room, Topic
-
 
 # rooms = [
 #     {"id": 1, "name": "Lets learn python!"},
 #     {"id": 2, "name": "Design with me !"},
 #     {"id": 3, "name": "Lets learn golang!"},
 # ]
+
+
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect("home")
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -58,6 +64,7 @@ def room(request, pk):
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -70,9 +77,13 @@ def create_room(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -84,8 +95,13 @@ def update_room(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
